Separate out server creation into module for tests
